+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/Row.kt
@@ -9,10 +9,14 @@ import ru.yandex.money.tools.grafana.dsl.json.jsonObject
  * @author Dmitry Komarov (komarovdmitry@yamoney.ru)
  * @since 7/21/18
  */
-class Row(private val basePanel: Panel) : Panel {
+class Row(
+    private val basePanel: Panel,
+    private val repeat: String?
+) : Panel {
 
     override fun toJson() = jsonObject(basePanel.toJson()) {
         "type" to "row"
+        "repeat" to repeat
         "collapsed" to false
         "panels" to emptyJsonArray()
     }