+++ src/test/kotlin/ru/yandex/money/tools/grafana/dsl/variables/VariablesTest.kt
@@ -61,4 +61,5 @@ class VariablesTest {
         variable.asVariable() shouldEqual "\$variable"
         variable.toJson().toString() shouldEqualToJson jsonFile("IntervalVar.json")
     }
+
 }